--- src/main/java/TaskType.java
@@ -1,6 +1,0 @@
-/**
- * Types of tasks that can be recorded to TaskList.
- */
-public enum TaskType {
-    Todo, Event, Deadline;
-}
