+++ src/main/java/stevie/StevieUi.java
@@ -1,3 +1,5 @@
+package stevie;
+
 import java.util.Scanner;
 
 public class StevieUi {
