+++ src/main/java/stevie/command/DeleteCommand.java
@@ -1,20 +1,20 @@
-public class MarkCommand extends Command {
-    private final boolean markDone;
+package stevie.command;
+
+import stevie.exception.StevieException;
+import stevie.StevieUi;
+import stevie.task.TaskDataHandler;
+import stevie.task.TaskList;
+
+public class DeleteCommand extends Command {
     private final int taskIdx;
-    public MarkCommand(boolean mark, int idx) {
-        markDone = mark;
+    public DeleteCommand(int idx) {
         taskIdx = idx;
     }
-
     @Override
     public boolean execute(TaskList tasks, TaskDataHandler storage, StevieUi ui) {
         String out;
         try {
-            if (markDone) {
-                out = tasks.markDone(taskIdx);
-            } else {
-                out = tasks.markUndone(taskIdx);
-            }
+            out = tasks.delete(taskIdx);
         } catch (StevieException ex) {
             ui.outputMessage(ex.getMessage());
             return false;
