+++ src/main/java/stevie/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package stevie.command;
+
+import stevie.StevieUi;
+import stevie.task.TaskDataHandler;
+import stevie.task.TaskList;
+
 public class ExitCommand extends Command {
     @Override
     public boolean execute(TaskList tasks, TaskDataHandler storage, StevieUi ui) {
