+++ src/main/java/stevie/exception/StevieException.java
@@ -1,5 +1,7 @@
+package stevie.exception;
+
 /**
- * The StevieException wraps all checked exceptions that is related to
+ * The stevie.exception.StevieException wraps all checked exceptions that is related to
  * erroneous user inputs.
  */
 public class StevieException extends Exception {
