+++ src/main/java/stevie/exception/TaskException.java
@@ -1,0 +1,13 @@
+package stevie.exception;
+
+import stevie.exception.StevieException;
+
+/**
+ * The stevie.exception.TaskException wraps all checked exceptions that is related to
+ * user inputs to the stevie.Stevie chat bot, that is impossible for stevie.task.TaskList to fulfil.
+ */
+public class TaskException extends StevieException {
+    public TaskException(String message) {
+        super(message);
+    }
+}
