+++ src/main/java/stevie/task/DeadlineTask.java
@@ -1,8 +1,9 @@
-import java.text.SimpleDateFormat;
+package stevie.task;
+
 import java.util.Date;
 
 /**
- * A type of task that can be characterised as an deadline. A <code>DeadlineTask</code> contains
+ * A type of task that can be characterised as an deadline. A <code>stevie.task.DeadlineTask</code> contains
  * a name and an deadline.
  */
 public class DeadlineTask extends Task {
