+++ src/main/java/stevie/task/TaskType.java
@@ -1,0 +1,8 @@
+package stevie.task;
+
+/**
+ * Types of tasks that can be recorded to stevie.task.TaskList.
+ */
+public enum TaskType {
+    Todo, Event, Deadline;
+}
